+++ CHANGELOG.md
@@ -153,6 +153,8 @@
 * New config option: CLICON_CLI_MODEL_TREENAME defining name of generated syntax tree if CLIXON_CLI_GENMODEL is set.
 
 ### Corrected Bugs
+* Did not check for missing keys in validate. [Key of a list isn't mandatory](https://github.com/clicon/clixon/issues/73)
+  * Problem here is that you can still add keys via netconf - since validation is a separate step, but in cli or restconf it should not be possible.
 * Partially corrected: [yang type range statement does not support multiple values](https://github.com/clicon/clixon/issues/59).
   * Should work for netconf and restconf, but not for CLI.
 * Fixed: [Range parsing is not RFC 7950 compliant](https://github.com/clicon/clixon/issues/71)